+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Check in diff of `package.json`/`package-lock.json` from moving to NPM v7
 - Update ssri dependency.
+- Externalize OpenLayers dependency
 
 ## [v1.4.2] - 2021-04-02